Move image components to components folder
